refactor: pricing 상수 분리
